⚗️ Add tests for `Inference` & `list_fine_tunes`
