+++ tests/test_utils.py
@@ -2,17 +2,9 @@ import json
 import tempfile
 
 from opentrain.utils import (
-    list_fine_tunes,
     prepare_openai_dataset,
     validate_openai_dataset,
 )
-
-
-def test_list_fine_tunes() -> None:
-    fine_tunes = list_fine_tunes(just_succeeded=True)
-    assert isinstance(fine_tunes, list)
-    assert len(fine_tunes) > 0
-    assert isinstance(fine_tunes[0], str)
 
 
 def test_prepare_openai_dataset() -> None:
